Fixed mapper issues and extended quat class

+++ [Library]/assetify_mapper/utilities/mapper.lua
@@ -50,6 +50,7 @@ local imports = {
     toJSON = toJSON,
     fromJSON = fromJSON,
     string = string,
+    Vector3 = Vector3,
     table = table,
     file = file,
     quat = quat,
@@ -98,10 +99,11 @@ end
 if localPlayer then
     mapper.axis = {
         validAxesTypes = {"slate", "ring"},
+        validLocationIndexes = {"posX", "posY", "posZ", "rotX", "rotY", "rotZ"},
         validAxes = {
-            x = {color = {255, 0, 0}, rotation = {slate = {0, 0, 90}, ring = {0, 0, 0}}},
-            y = {color = {0, 255, 0}, rotation = {slate = {0, 0, 0}, ring = {0, 90, 0}}},
-            z = {color = {0, 0, 255}, rotation = {slate = {90, 0, 0}, ring = {90, 0, 0}}}
+            x = {color = {255, 0, 0}, rotation = {slate = {0, 0, 90}, ring = {0, 90, 0}}},
+            y = {color = {0, 255, 0}, rotation = {slate = {0, 0, 0}, ring = {90, 0, 0}}},
+            z = {color = {0, 0, 255}, rotation = {slate = {90, 0, 0}, ring = {0, 0, 0}}}
         }
     }
     mapper.cache = {keys = {}}
@@ -209,15 +211,15 @@ if localPlayer then
             mapper.ui.toolWnd.renderUI()
             mapper.ui.notif.renderUI()
             mapper.translationMode = (mapper.isTargettingDummy and (mapper.translationMode or {})) or false
-            local isCursorTranslation = false
+            local translationIndex, isCursorTranslation = false, false
             if mapper.translationMode then
                 mapper.translationMode.element = mapper.isTargettingDummy
                 mapper.translationMode.type = mapper.translationMode.type or 1
                 mapper.translationMode.axis = mapper.translationMode.axis or "x"
-                mapper.translationMode.posX, mapper.translationMode.posY, mapper.translationMode.posZ, mapper.translationMode.rotX, mapper.translationMode.rotY, mapper.translationMode.rotZ = imports.getElementLocation(mapper.translationMode.element)
+                mapper.translationMode.posX, mapper.translationMode.posY, mapper.translationMode.posZ, mapper.translationMode.rotX, mapper.translationMode.rotY, mapper.translationMode.rotZ = imports.getElementLocation(mapper.translationMode.element, "ZYX")
                 if not CLIENT_MTA_WINDOW_ACTIVE then
                     local isSlateTranslation = mapper.axis.validAxesTypes[(mapper.translationMode.type)] == "slate"
-                    local translationIndex = ((mapper.translationMode.axis == "x") and ((isSlateTranslation and "posX") or "rotX")) or ((mapper.translationMode.axis == "y") and ((isSlateTranslation and "posY") or "rotY")) or ((mapper.translationMode.axis == "z") and ((isSlateTranslation and "posZ") or "rotZ")) or false
+                    translationIndex = ((mapper.translationMode.axis == "x") and ((isSlateTranslation and "posX") or "rotX")) or ((mapper.translationMode.axis == "y") and ((isSlateTranslation and "posY") or "rotY")) or ((mapper.translationMode.axis == "z") and ((isSlateTranslation and "posZ") or "rotZ")) or false
                     if translationIndex then
                         local translationSpeed, translationValue = (imports.getKeyState(availableControls.speedUp) and availableControlSpeeds.fast) or (imports.getKeyState(availableControls.speedDown) and availableControlSpeeds.slow) or availableControlSpeeds.normal, nil
                         if camera.isCursorVisible then
@@ -231,27 +233,53 @@ if localPlayer then
                                 else
                                     translationValue = (translationValue*360)%360
                                 end
-                            end
+                            end                            
                         end
+                        local __translationIndex = "__"..translationIndex
                         if not isCursorTranslation then
                             translationValue = ((imports.getKeyState(availableControls.valueUp) and translationSpeed) or (imports.getKeyState(availableControls.valueDown) and -translationSpeed) or 0)
-                            if isSlateTranslation then translationValue = translationValue*0.75 end
-                            mapper.translationMode[translationIndex] = mapper.translationMode[translationIndex] + translationValue
-                        else
-                            mapper.translationMode[("__"..translationIndex)] = mapper.translationMode[("__"..translationIndex)] or {
-                                native = mapper.translationMode[translationIndex],
-                                previous = 0, current = 0, offset = translationValue,
-                                speed = translationSpeed
-                            }
-                            if mapper.translationMode[("__"..translationIndex)].speed ~= translationSpeed then
-                                mapper.translationMode[("__"..translationIndex)].speed = translationSpeed
-                                mapper.translationMode[("__"..translationIndex)].previous = mapper.translationMode[("__"..translationIndex)].previous + mapper.translationMode[("__"..translationIndex)].current
-                                mapper.translationMode[("__"..translationIndex)].offset = translationValue
+                            if isSlateTranslation then
+                                translationValue = translationValue*0.75
+                                mapper.translationMode[translationIndex] = mapper.translationMode[translationIndex] + translationValue
+                            else
+                                mapper.translationMode[translationIndex] = translationValue
                             end
-                            mapper.translationMode[("__"..translationIndex)].current = (translationValue - mapper.translationMode[("__"..translationIndex)].offset)*translationSpeed
-                            mapper.translationMode[translationIndex] = mapper.translationMode[("__"..translationIndex)].native + mapper.translationMode[("__"..translationIndex)].previous + mapper.translationMode[("__"..translationIndex)].current
+                            mapper.translationMode[__translationIndex] = nil
+                        else
+                            local __translationIndex = "__"..translationIndex
+                            mapper.translationMode[__translationIndex] = mapper.translationMode[__translationIndex] or {
+                                native = mapper.translationMode[translationIndex]
+                            }
+                            if mapper.translationMode[__translationIndex].speed ~= translationSpeed then
+                                mapper.translationMode[__translationIndex].speed = translationSpeed
+                                mapper.translationMode[__translationIndex].previous = mapper.translationMode[__translationIndex].current or 0
+                                mapper.translationMode[__translationIndex].offset = translationValue
+                            end
+                            mapper.translationMode[__translationIndex].__current = mapper.translationMode[__translationIndex].current or 0
+                            mapper.translationMode[__translationIndex].current = mapper.translationMode[__translationIndex].previous + (translationValue - mapper.translationMode[__translationIndex].offset)*translationSpeed
+                            if isSlateTranslation then
+                                mapper.translationMode[translationIndex] = mapper.translationMode[__translationIndex].native + mapper.translationMode[__translationIndex].current
+                            else
+                                mapper.translationMode[translationIndex] = mapper.translationMode[__translationIndex].current - mapper.translationMode[__translationIndex].__current
+                            end
                         end
-                        imports.setElementLocation(mapper.translationMode.element, mapper.translationMode.posX, mapper.translationMode.posY, mapper.translationMode.posZ, mapper.translationMode.rotX, mapper.translationMode.rotY, mapper.translationMode.rotZ)
+                        if not isSlateTranslation then
+                            local translationValues = {}
+                            for i = 4, #mapper.axis.validLocationIndexes, 1 do
+                                local j = mapper.axis.validLocationIndexes[i]
+                                if translationIndex == j then
+                                    translationValues[j] = mapper.translationMode[j]
+                                else
+                                    translationValues[j] = 0
+                                end
+                            end
+                            local _, _, _, __rotX, __rotY, __rotZ = imports.getElementLocation(mapper.translationMode.element, "ZYX")
+                            local rotQuat = imports.quat.new(imports.quat.fromEuler(__rotX, __rotY, __rotZ))
+                            local __rotQuat = imports.quat.fromVectorAngle(imports.Vector3(1, 0, 0), translationValues.rotX)*imports.quat.fromVectorAngle(imports.Vector3(0, 1, 0), translationValues.rotY)*imports.quat.fromVectorAngle(imports.Vector3(0, 0, 1), translationValues.rotZ) 
+                            rotQuat = __rotQuat*rotQuat
+                            mapper.translationMode.rotX, mapper.translationMode.rotY, mapper.translationMode.rotZ = imports.quat.toEuler(rotQuat[1], rotQuat[2], rotQuat[3], rotQuat[4])
+                        end
+                        imports.setElementLocation(mapper.translationMode.element, mapper.translationMode.posX, mapper.translationMode.posY, mapper.translationMode.posZ, mapper.translationMode.rotX, mapper.translationMode.rotY, mapper.translationMode.rotZ, "ZYX")
                     end
                 end
             end
@@ -270,10 +298,12 @@ if localPlayer then
                     imports.setElementCollisionsEnabled(mapper.axis[j][k].LODInstance, isCollisionEnabled)
                 end
             end
-            mapper.isCursorTranslation = isCursorTranslation
-            mapper.cache.prevCursorY = (mapper.isCursorTranslation and mapper.cache.prevCursorY) or false
-            if mapper.translationMode and not mapper.isCursorTranslation then
-                mapper.translationMode.__posX, mapper.translationMode.__posY, mapper.translationMode.__posZ, mapper.translationMode.__rotX, mapper.translationMode.__rotY, mapper.translationMode.__rotZ = nil, nil, nil, nil, nil, nil
+            mapper.cache.prevCursorY = (isCursorTranslation and mapper.cache.prevCursorY) or false
+            if mapper.translationMode and not translationIndex then
+                for i = 1, #mapper.axis.validLocationIndexes, 1 do
+                    local j = mapper.axis.validLocationIndexes[i]
+                    mapper.translationMode[("__"..j)] = nil
+                end
             end
         end
     end
@@ -462,7 +492,9 @@ else
         imports.table.remove(mapper.rwAssets[(mapper.cacheManifestPath)], sceneIndex)
         imports.file.write(mapper.cacheManifestPath, imports.toJSON(mapper.rwAssets[(mapper.cacheManifestPath)]))
         local sceneIPLPath = sceneCache.."scene.ipl"
-        imports.file.delete(sceneIPLPath)
+        if imports.file.exists(sceneIPLPath) then
+            imports.file.delete(sceneIPLPath)
+        end
         mapper.syncCacheManifest(source)
         imports.triggerClientEvent(source, "Assetify:Mapper:onNotification", source, "Scene successfully deleted. ["..sceneIPLPath.."]", availableColors.success)
     end)
